feat:쿠폰 Swagger API 적용
