+++ coupon-service/src/main/resources/application.yml
@@ -9,3 +9,5 @@ spring:
     include:
       - datasource
       - zipkin
+      - eureka
+      - swagger
